aidl: Set PMF as optional when framework doesn't set it
With cross AKM roaming changes framework is not setting PMF as
required for multi-AKM connect case so PMF value in SSID block
is false although SAE AKM is set. This is causing association
request go without setting MFPC bit set when connecting to WPA3
AP. To avoid this always set PMF as optional in ssid profile when
wifi frameworks doesn't set it.

CRs-Fixed: 3367089
Change-Id: I6eda6cfb1a87ba1c00995ee0a42f416d6d24ed87