eap_proxy: During deinit unblock the join based on PPID
When the wpa_supplicant is started using ctrlapp_dut, it uses -B
flag in the command. And this flag daemonizes the wpa_supplicant.
The test case then tries to kill the wpa_supplicant and
reinitialize it.
Initially when the wpa_supplicant was started, eap_proxy gets
created as a child thread under this parent. And when the
wpa_supplicant gets daemonized and created as a new process, the
eap_proxy thread also dies along with its parent wpa_supplicant.
But as the dead thread ID is passed in the pthread_join call it
waits for an indefinite time. And this block the wpa_supplicant
termination, results in wpa_supplicant restart issue.

To fix this compare the PPID of the wpa_supplicant init and
deinit to unblock the pthread_join during deinit process.
As the PPID of wpa_supplicant before daemonize will be different
than the PPID after daemonize because the wpa_supplicant will be
adopted by init process which has a PID '1'.

Change-Id: I74ed4072ef3eb53af7b93d8838cfd530ea7532aa
CRs-Fixed: 3393888