Update PMK in wpa_sm when roam+auth event indicated with authorized flag
Currently, the PMK used by the driver is not updated to wpa_sm when
roaming is completed by the driver with the cached PMKSA and the
roam+auth event is indicated with the authorized flag.

To fix this, identify the PMKSA entry from the PMKID sent in
Reassociation Request frame and update the correct PMK to wpa_sm from
the PMKSA entry.

Signed-off-by: Veerendranath Jakkam <quic_vjakkam@quicinc.com>

CRs-Fixed: 3399582
Git-repo: git://w1.fi/hostap.git
Git-commit: b2bf7e39ecc9f52e7a0f11a9f8d9e4fc2ce57bd6
Change-Id: Ief24ed6d0a130f4027755629a704f9e67ef36428